u64

+++ euler92/src/main.rs
@@ -8,7 +8,7 @@ fn main() {
     println!("{}", count);
 }
 
-fn square_chain_arrive(n: u32) -> u32 {
+fn square_chain_arrive(n: u64) -> u64 {
     let mut x = n;
     while x != 1 && x != 89 {
         x = square_sum(x);
@@ -16,9 +16,9 @@ fn square_chain_arrive(n: u32) -> u32 {
     x
 }
 
-fn square_sum(n: u32) -> u32 {
+fn square_sum(n: u64) -> u64 {
     n.to_string()
         .chars()
-        .map(|x| x.to_digit(10).unwrap().pow(2))
-        .sum::<u32>()
+        .map(|x| x.to_digit(10).unwrap().pow(2) as u64)
+        .sum::<u64>()
 }
